+++ lib/passfort/errors.rb
@@ -7,6 +7,7 @@ require "passfort/errors/invalid_api_key_error"
 require "passfort/errors/invalid_input_data_error"
 require "passfort/errors/request_error"
 require "passfort/errors/unparseable_response_error"
+require "passfort/errors/timeout_error"
 
 module Passfort
   module Errors