+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Changelog
 Unreleased
 ----------
 
-No changes.
+- Raise `Passfort::Errors::TimeoutError` when a request times out.
 
 0.2.2
 ----------